fix: 【5.7.0.13】特定文档：截图与录屏 UE FLOW文档，文档信息界面的缩略图未显示
【5.7.0.13】特定文档：截图与录屏 UE FLOW文档，文档信息界面的缩略图未显示

Log: 【5.7.0.13】特定文档：截图与录屏 UE FLOW文档，文档信息界面的缩略图未显示

link: https://pms.uniontech.com/zentao/bug-view-42394.html

+++ src/document/PDFModel.cpp
@@ -429,8 +429,8 @@ QImage PDFPage::render(Dr::Rotation rotation, const double scaleFactor, const QR
 
 QImage PDFPage::render(int width, int height, Qt::AspectRatioMode) const
 {
-    int horizontalResolution = 72 * width / m_page->pageSize().width();
-    int verticalResolution = 72 * height / m_page->pageSize().height();
+    qreal horizontalResolution = 72.0 * width / m_page->pageSize().width();
+    qreal verticalResolution = 72.0 * height / m_page->pageSize().height();
 
     return render(horizontalResolution, verticalResolution, Dr::RotateBy0, QRect());
 }
